Change field type

+++ app_run/models.py
@@ -17,7 +17,7 @@ class Run(models.Model):
 
 class AthleteInfo(models.Model):
     goals = models.CharField(max_length=200, null=True)
-    weight = models.DecimalField(max_digits=5, decimal_places=2, null=True)
+    weight = models.IntegerField(null=True)
     user_id = models.OneToOneField(
         User,
         on_delete=models.CASCADE,
